requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
langchain-google-genai
google-generativeai
sentence-transformers
torch
azure-search-documents
azure-core
azure-identity
langchain-openai